+++ plat/allwinner/common/sunxi_topology.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2018, ARM Limited and Contributors. All rights reserved.
+ * Copyright (c) 2017-2019, ARM Limited and Contributors. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -30,7 +30,7 @@ int plat_core_pos_by_mpidr(u_register_t mpidr)
 		return -1;
 	}
 
-	return cluster * PLATFORM_MAX_CPUS_PER_CLUSTER + core;
+	return (int) cluster * PLATFORM_MAX_CPUS_PER_CLUSTER + core;
 }
 
 const unsigned char *plat_get_power_domain_tree_desc(void)
